Refactor: extracted some methods #4
